requirements.txt
flask==3.0.0
anthropic==0.40.0
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.0
email-validator==2.1.0
